requirements.txt
build
xmlsec
python-jose
